Fix test_json tests with Lazy Imports enabled
Summary:

When `Lib/test/test_json/__init__.py` imports cjson and pyjson
using `import_helper.import_fresh_module`, but just before returning
the loaded module, it restores the previous state of `sys.modules`;
in doing so, all still lazy objects in the "fresh" module being returned
will attempt to get the objects from the modules in the newly restored
`sys.modules`, obtaining a different version of the module to that which
was supposed to be there.

To fix this problem, this commit instead makes sure all lazy objects in
the returned "fresh" module are resolved, before to returning it.

Test Plan:

```
./python.exe -L -m test -v test_json
```